use urfave/cli to handle CLI
